Update initialisation script to add symlinks.
With WD MyCloud OS3 version 2.31.163 there are two additional symlinks named "Firefly" and perl5.10". Addition is to include these two symlinks which are omitted.  They are missing in /usr/local/modules/opt but present on my NAS,  Maybe because of other applications installed on WD's app store?  The problem is that entware's init script needs to look at the contents of the original /opt directory and copy the contents to /shares/Volume_1/entware/ upon the NAS's start-up.

My suggestion to this script shows where the change needs to be processed.